#1356 Fix integer custom field

+++ thehive/app/org/thp/thehive/models/Alert.scala
@@ -11,14 +11,14 @@ case class AlertCustomField(
     order: Option[Int] = None,
     stringValue: Option[String] = None,
     booleanValue: Option[Boolean] = None,
-    integerValue: Option[Int] = None,
+    integerValue: Option[Long] = None,
     floatValue: Option[Double] = None,
     dateValue: Option[Date] = None
 ) extends CustomFieldValue[AlertCustomField] {
   override def order_=(value: Option[Int]): AlertCustomField            = copy(order = value)
   override def stringValue_=(value: Option[String]): AlertCustomField   = copy(stringValue = value)
   override def booleanValue_=(value: Option[Boolean]): AlertCustomField = copy(booleanValue = value)
-  override def integerValue_=(value: Option[Int]): AlertCustomField     = copy(integerValue = value)
+  override def integerValue_=(value: Option[Long]): AlertCustomField    = copy(integerValue = value)
   override def floatValue_=(value: Option[Double]): AlertCustomField    = copy(floatValue = value)
   override def dateValue_=(value: Option[Date]): AlertCustomField       = copy(dateValue = value)
 }
